requirements.txt
narrative-graph-representations==1.0
numpy==1.24.2
pandas==1.5.3
PuLP==2.7.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sentence-transformers==2.2.2
rdflib==6.3.2
bertopic==0.15.0
pybind11==2.10.4
hdt==2.3